refactor(#20): EncryptedSharedPreferencesManager -> DataStoreManager
